backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
transformers==4.40.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
aiosqlite==0.20.0
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.30.1
scipy==1.13.0
